requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.5.0
djangorestframework==3.12.1
gunicorn==19.7.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==3.3.1
wrapt==1.12.1
# psycopg2==2.7.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
django-allauth==0.39.1
django-rest-auth==0.9.5
idna==2.10
oauthlib==3.0.1
python3-openid==3.2.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10